Assignment 07 files in a compressed file format
Three python script files. One showing handling of exceptions. One showing basic Pickling, and a third script that integrates lesson 6 functionality with the exception handling and pickling ( uses a menu function to help you view output that is generated from my newly learned Python skills). The link to my Blog is in a text file within the compressed file. 

This is the Blog URL
https://www.blogger.com/blogger.g?blogID=2148431566826702773#editor/target=post;postID=4640133620741494092;onPublishedMenu=allposts;onClosedMenu=allposts;postNum=0;src=postname